--- src/main/java/com/toy/namoner/common/web/WebConfig.java
@@ -1,30 +1,0 @@
-package com.toy.namoner.common.web;
-
-import com.toy.namoner.common.logs.LoggingInterceptor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new LoggingInterceptor())
-                .order(1)
-                .addPathPatterns("/**");
-    }
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        //todo 정식 배포 시 수정 필요
-        registry.addMapping("/**")
-                .allowedOriginPatterns("*")
-                .allowedMethods("GET", "POST", "PUT", "DELETE")
-                .allowedHeaders("Authorization", "Content-Type")
-                .exposedHeaders("Custom-Header")
-                .allowCredentials(true)
-                .maxAge(3600);
-    }
-}
