--- src/main/java/com/toy/namoner/config/S3ImageServiceImpl.java
@@ -1,76 +1,0 @@
-package com.toy.namoner.config;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class S3ImageServiceImpl implements ImageService {
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-    @Value("${cloud.aws.region.static}")
-    private String region;
-
-    private final AmazonS3 amazonS3;
-
-    @Override
-    public List<String> uploadFiles(String prefix, List<MultipartFile> multipartFiles) {
-        ArrayList<String> fileNameList = new ArrayList<>();
-
-        multipartFiles.forEach(file -> {
-            String fileName = uploadFile(prefix, file);
-            fileNameList.add(fileName);
-        });
-
-        return fileNameList;
-    }
-
-    @Override
-    public String uploadFile(String prefix, MultipartFile multipartFile) {
-        String fileName = prefix + "/" + createFileName(multipartFile.getOriginalFilename());
-        ObjectMetadata objectMetadata = new ObjectMetadata();
-        objectMetadata.setContentLength(multipartFile.getSize());
-        objectMetadata.setContentType(multipartFile.getContentType());
-
-        try (InputStream inputStream = multipartFile.getInputStream()) {
-            amazonS3.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
-                    .withCannedAcl(CannedAccessControlList.PublicRead));
-        } catch (IOException e) {
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
-        }
-        return fileName;
-    }
-
-    @Override
-    public String getFileUrl(String fileName) {
-        if (fileName == null || fileName.isEmpty()) {
-            return null;
-        }
-        return "https://" + bucket + ".s3." + region + ".amazonaws.com/" + fileName;
-    }
-    private String createFileName(String fileName) {
-        return UUID.randomUUID().toString().concat(getFileExtension(fileName));
-    }
-
-    private String getFileExtension(String fileName) {
-        try {
-            return fileName.substring(fileName.lastIndexOf("."));
-        } catch (StringIndexOutOfBoundsException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 형식의 파일" + fileName + ") 입니다.");
-        }
-    }
-}
