+++ src/main/java/com/toy/namoner/domain/letter/model/enums/FontType.java
@@ -1,0 +1,5 @@
+package com.toy.namoner.domain.letter.model.enums;
+
+public enum FontType {
+    Pretendard_R, Gmarket_R, Leeseoyun, Gowun_R, KCC, Kangwon,
+}
