+++ src/main/java/com/toy/namoner/domain/letter/model/enums/LetterPaperType.java
@@ -1,0 +1,11 @@
+package com.toy.namoner.domain.letter.model.enums;
+
+public enum LetterPaperType {
+    GRAPH_PAPER("모눈종이"), BASIC_NOTE("유선노트"), PHOTO_POSTCARD("포토엽서"), POLAROID("폴라로이드"),
+    ;
+
+    private String description;
+    LetterPaperType(String description) {
+
+    }
+}
