+++ src/main/java/com/toy/namoner/domain/letter/model/enums/LetterType.java
@@ -1,0 +1,5 @@
+package com.toy.namoner.domain.letter.model.enums;
+
+public enum LetterType {
+    NORMAL, RESERVED
+}
