+++ src/main/java/com/toy/namoner/domain/letter/repository/LetterRepository.java
@@ -1,0 +1,8 @@
+package com.toy.namoner.domain.letter.repository;
+
+import com.toy.namoner.domain.letter.model.Letter;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LetterRepository extends JpaRepository<Letter, String> {
+
+}
