+++ src/main/java/com/toy/namoner/domain/user/model/enums/UserStatus.java
@@ -1,0 +1,5 @@
+package com.toy.namoner.domain.user.model.enums;
+
+public enum UserStatus {
+    NOT_SIGNED, SIGNED
+}
