--- src/main/java/com/toy/namoner/letter/domain/enums/FontType.java
@@ -1,5 +1,0 @@
-package com.toy.namoner.letter.domain.enums;
-
-public enum FontType {
-    Pretendard_R, Gmarket_R, Leeseoyun, Gowun_R, KCC, Kangwon,
-}
