--- src/main/java/com/toy/namoner/letter/domain/enums/LetterPaperType.java
@@ -1,11 +1,0 @@
-package com.toy.namoner.letter.domain.enums;
-
-public enum LetterPaperType {
-    GRAPH_PAPER("모눈종이"), BASIC_NOTE("유선노트"), PHOTO_POSTCARD("포토엽서"), POLAROID("폴라로이드"),
-    ;
-
-    private String description;
-    LetterPaperType(String description) {
-
-    }
-}
