--- src/main/java/com/toy/namoner/letter/domain/enums/LetterType.java
@@ -1,5 +1,0 @@
-package com.toy.namoner.letter.domain.enums;
-
-public enum LetterType {
-    NORMAL, RESERVED
-}
