--- src/main/java/com/toy/namoner/letter/repository/LetterRepository.java
@@ -1,8 +1,0 @@
-package com.toy.namoner.letter.repository;
-
-import com.toy.namoner.letter.domain.Letter;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LetterRepository extends JpaRepository<Letter, String> {
-
-}
