--- src/main/java/com/toy/namoner/user/domain/enums/UserStatus.java
@@ -1,5 +1,0 @@
-package com.toy.namoner.user.domain.enums;
-
-public enum UserStatus {
-    NOT_SIGNED, SIGNED
-}
