--- src/main/java/com/toy/namoner/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.toy.namoner.user.repository;
-
-import com.toy.namoner.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, String> {
-
-    Optional<User> findByPhone(String phone);
-
-}
